pyproject.toml
[project]
name = "pyxtxt"
version = "0.3.5"
description = "A Python library for extracting text from different types of files (PDF, DOCX, PPTX, XLSX, ODT, ecc.)."
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Topic :: Text Processing",
  "Topic :: Utilities"
]

readme = "README.md"
requires-python = ">=3.7"

license = { file = "LICENSE" }
authors = [
  { name = "Giuseppe Levi", email = "giuseppe.levi@gmail.com" }
]
dependencies = [
  "python-magic; sys_platform != 'win32'",
  "python-magic-bin; sys_platform == 'win32'"
]
[project.optional-dependencies]
pdf = [
  "PyMuPDF",
]
docx = [
  "python-docx",
]
presentation = [
  "python-pptx",
]
spreadsheet = [
  "openpyxl",
  "xlrd",
]
odf = [
  "odfpy",
]
html = [
  "beautifulsoup4",
  "lxml",
]
doc = [
]
markdown = [
  "markdown",
  "beautifulsoup4",
]
epub = [
  "ebooklib", 
  "beautifulsoup4",
]
rtf = [
  "striprtf",
]
email = [
  "beautifulsoup4",
]
outlook = [
  "extract-msg",
  "beautifulsoup4",
]
latex = [
  "pylatexenc",
]
audio = [
  "openai-whisper",
]
ocr = [
  "easyocr",
  "pillow",
]
ocr-ollama = [
  "ollama",
  "pillow",
]
all = [
  "PyMuPDF",
  "python-docx",
  "python-pptx",
  "openpyxl",
  "xlrd",
  "odfpy",
  "beautifulsoup4",
  "lxml",
  "markdown",
  "ebooklib",
  "striprtf",
  "extract-msg",
  "pylatexenc",
  "openai-whisper",
  "easyocr",
  "pillow",
  "ollama",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"


[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
"*" = ["examples.py"]

